Split out redirects into their own file
