+++ inlang/source-code/paraglide/paraglide-next/src/app/createI18n.client.ts
@@ -1,7 +1,8 @@
 import { createLink } from "./Link"
 import { getLanguage } from "./getLanguage.client"
 import { availableLanguageTags, sourceLanguageTag } from "$paraglide/runtime.js"
-import { createNavigation, createRedirects } from "./navigation"
+import { createNavigation } from "./navigation"
+import { createRedirects } from "./redirect"
 import { createExclude } from "./exclude"
 import { createMiddleware } from "./middleware"
 import { I18nUserConfig, ResolvedI18nConfig } from "./config"
