+++ inlang/source-code/paraglide/paraglide-next/src/app/createI18n.server.ts
@@ -1,7 +1,8 @@
 import { createLink } from "./Link"
 import { getLanguage } from "./getLanguage.server"
 import { availableLanguageTags, sourceLanguageTag } from "$paraglide/runtime.js"
-import { createNoopNavigation, createRedirects } from "./navigation"
+import { createNoopNavigation } from "./navigation"
+import { createRedirects } from "./redirect"
 import { createExclude } from "./exclude"
 import { createMiddleware } from "./middleware"
 import { resolveUserPathDefinitions } from "@inlang/paraglide-js/internal/adapter-utils"
